Update the stream
A stream is a collection of nodes. The nodes are created by the stream
from a string or an array of strings. The stream will feed you the next
node via #next; a StreamIsEmpty will be raised when you reach the end of
the stream. The stream is NOT rewindable.